requirements/base.txt
django==5.1.2
pillow==10.4.0
djangorestframework==3.15.2
python-decouple==3.8
psycopg==3.2.3
celery==5.4.0
django-mptt==0.16.0
drf-nested-routers==0.94.1
drf-spectacular==0.27.2